+++ src/util/mod.rs
@@ -7,3 +7,7 @@ pub use self::opt_uc_x_param::OptUcXParamParseError;
 mod opt_color_when;
 pub use self::opt_color_when::OptColorWhen;
 pub use self::opt_color_when::OptColorWhenParseError;
+
+mod opt_around_num;
+pub use self::opt_around_num::OptAroundNum;
+pub use self::opt_around_num::OptAroundNumParseError;